Тест с полной матрицей и с RT и IT, равными 32 бита каждый.

+++ src/prec.jl
@@ -9,6 +9,6 @@ export IT, RT
 IT = Int32
 
 # Определение типа вещественных чисел
-RT = Float64
+RT = Float32
 
 end
